seperating windows and unix platform to make seperate build
